Pedidos

+++ src/app/pedidos/car-pedido/car-pedido.component.ts
@@ -6,6 +6,7 @@ import { MatTableModule } from '@angular/material/table';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 import { ApiService } from '../../api/api.service';
+import  Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-car-pedido',
@@ -24,18 +25,35 @@ displayedColumns=[
   'total'
 ];
 delete(element) {
-  let idx= this.dataSource.indexOf(element);
-  
-  this.dataSource.splice(idx,1)
-  
-  this.dataSource=[...this.dataSource]
-  this.lsts.setItem('pedido',this.dataSource);
-  this.socket.test(this.dataSource.length)
-  
-  
 
-  
-  
+  Swal.fire({
+    title: 'Confirmar',
+    text: 'Seguro que deseas quitar el producto del pedido?',
+    icon: 'warning',
+    showCancelButton:true,
+    confirmButtonText: 'Si',
+    cancelButtonText:'No',
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+  }).then((result)=>{
+    if(result.isConfirmed){
+      let idx= this.dataSource.indexOf(element);
+      
+      this.dataSource.splice(idx,1)
+      
+      this.dataSource=[...this.dataSource]
+      this.lsts.setItem('pedido',this.dataSource);
+      this.socket.test(this.dataSource.length)
+      Swal.fire({
+        title: "Deleted!",
+        text: "Your file has been deleted.",
+        icon: "success"
+      });
+    }else{
+      result.dismiss === Swal.DismissReason.cancel
+    }
+  })
+
 
 }
